+++ src/app/page.tsx
@@ -1,12 +1,11 @@
-import { MainNav } from '@/components/MainNav';
-import { TransposeNote } from '@/components/TransposeNote/TransposeNote';
+import { TransposeMessage } from '@/components/TransposeMessage';
 
 import styles from "./page.module.css";
 
 export default function Home() {
   return (
     <main className={styles.main}>
-      <TransposeNote />
+      <TransposeMessage />
     </main>
   );
 }
