+++ src/components/Note/Note.module.css
@@ -1,14 +1,3 @@
-.message {
-  font-size: 0.875rem;
-  text-align: center;
-}
-
-@media (width >= 45rem) {
-  .message {
-    font-size: 1.5rem;
-  }
-}
-
 .accidental {
   font-size: 0.625em;
   inset-block-end: 0.25em;
